Display witness script in txin script, fix lots of bugs

+++ BlockchainSQL.Web/Code/IBlockchainRepository.cs
@@ -26,7 +26,7 @@ namespace BlockchainSQL.Web.Code
 
         Task<IEnumerable<StatementLine>> GetStatementLines(string address);
 
-        Task<ScriptSummary> GetScriptSummary(int scriptId);
+        Task<ScriptSummary> GetScriptSummary(long scriptId);
 
         Task<SearchResult> SearchHash(string text);
